+++ src/components/SearchInputModal.vue
@@ -6,12 +6,17 @@
         class="search-input"
         v-model="inputKeyword"
         ref="searchInput"
-        @keyup.enter="searchKeyword"
+        @focus="handleFocus"
+        @blur="handleBlur"
+        @input="handleInput"
+        @keyup.enter="throttleSearch"
       />
+      <span ref="customCursor" class="custom-cursor" v-show="isFocused"></span>
+      <span ref="measureSpan" class="measure-span">{{ inputKeyword }}</span>
       <button
         type="button"
         class="btn bg-violet-300 hover:bg-violet-400"
-        @click="searchKeyword"
+        @click="throttleSearch"
       >
         검색
       </button>
@@ -22,6 +27,8 @@
 <script setup lang="ts">
 import { ref, defineModel, watch, nextTick, onUpdated } from 'vue';
 import { css } from '@/utils/utils';
+
+import { throttle } from 'lodash-es';
 
 const showSearch = defineModel<boolean>('showSearch', { required: true });
 
@@ -42,6 +49,7 @@ const searchKeyword = async () => {
   emit('closeSearchModal', 'search', inputKeyword.value);
   resetData();
 };
+const throttleSearch = throttle(searchKeyword, 300);
 
 const dialogState = ref(false);
 const close = () => {
@@ -62,9 +70,34 @@ watch(showSearch, async (val) => {
   setTimeout(() => {
     if (val) {
       searchInput.value!.focus();
+      updateCursorPosition();
     }
   }, 100);
 });
+
+const isFocused = ref(false);
+const handleInput = () => {
+  updateCursorPosition();
+};
+const handleFocus = () => {
+  isFocused.value = true;
+};
+const handleBlur = () => {
+  isFocused.value = false;
+};
+
+const measureSpan = ref<HTMLSpanElement>();
+const customCursor = ref<HTMLSpanElement>();
+const updateCursorPosition = async () => {
+  await nextTick();
+
+  if (isFocused.value && measureSpan.value && customCursor.value) {
+    const textWidth = measureSpan.value.offsetWidth;
+    // input 패딩(5px) + 테두리 고려
+    // 200은 input의 너비
+    customCursor.value.style.left = `${textWidth + 5 - 200}px`; // 패딩/테두리 보정값
+  }
+};
 
 onUpdated(async () => {
   await nextTick();
@@ -92,6 +125,30 @@ onUpdated(async () => {
   border-radius: 5px;
   background-color: #fcf8e8;
   padding: 0 5px;
+  font-family: 'Monaco';
+  caret-color: transparent;
+}
+.custom-cursor {
+  display: none;
+  position: relative;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 2px;
+  height: 1.5rem;
+  background: #007bff;
+}
+.search-input:focus + .custom-cursor {
+  display: inline-block;
+  animation: blink 0.7s infinite;
+  caret-color: transparent; /* 기존 커서 숨기기 */
+}
+.measure-span {
+  position: absolute;
+  visibility: hidden;
+  white-space: pre; /* 공백 유지 */
+  font-size: 1.2rem;
+  font-family: 'Monaco';
+  padding: 0;
 }
 .btn {
   pointer-events: all;
@@ -101,5 +158,14 @@ onUpdated(async () => {
   margin-left: 5px;
   cursor: pointer;
   border-radius: 3px;
+}
+@keyframes blink {
+  0%,
+  100% {
+    opacity: 1;
+  }
+  50% {
+    opacity: 0;
+  }
 }
 </style>